+++ src/editor/exposenode.ts
@@ -83,7 +83,6 @@ export function init(n: HassNodeProperties) {
                 renderAlert('0.5.0');
                 break;
             case NodeType.Entity:
-            case NodeType.Webhook:
                 renderAlert();
                 break;
             case NodeType.Number:
@@ -95,6 +94,16 @@ export function init(n: HassNodeProperties) {
             case NodeType.Select:
                 if ($('#node-input-entityConfig').val() !== '_ADD_') {
                     renderAlert('1.4.0');
+                }
+                break;
+            case NodeType.Sentence:
+                if ($('#node-input-server').val() !== '_ADD_') {
+                    renderAlert('1.5.0');
+                }
+                break;
+            case NodeType.Webhook:
+                if ($('#node-input-server').val() !== '_ADD_') {
+                    renderAlert('1.6.0');
                 }
                 break;
             default:
@@ -119,7 +128,6 @@ function render() {
             }
             break;
         case NodeType.Entity:
-        case NodeType.Webhook:
             renderAlert();
             break;
         case NodeType.Number:
@@ -133,6 +141,17 @@ function render() {
                 renderAlert('1.4.0');
             }
             break;
+        case NodeType.Sentence:
+            if ($('#node-input-server').val() !== '_ADD_') {
+                renderAlert('1.5.0');
+            }
+            break;
+        case NodeType.Webhook:
+            if ($('#node-input-server').val() !== '_ADD_') {
+                renderAlert('1.6.0');
+            }
+            break;
+
         default:
             renderEventNode();
     }