+++ src/nodes/switch/SwitchController.ts
@@ -96,7 +96,7 @@ export default class SwitchController extends InputOutputController<
         const message: NodeMessage = {
             topic: 'triggered',
         };
-        if (data.payload) {
+        if (data.payload !== undefined) {
             message.payload = data.payload;
         }