+++ test/migrations/webhook.test.js
@@ -47,6 +47,14 @@ const VERSION_1 = {
     headersLocation: undefined,
     headersLocationType: undefined,
 };
+const VERSION_2 = {
+    ...VERSION_1,
+    version: 2,
+    method_post: true,
+    method_put: true,
+    method_get: false,
+    method_head: false,
+};
 
 describe('Migrations - Webhook Node', function () {
     describe('Version 0', function () {
@@ -63,8 +71,15 @@ describe('Migrations - Webhook Node', function () {
             expect(migratedSchema).to.eql(VERSION_1);
         });
     });
+    describe('Version 2', function () {
+        it('should update version 1 to version 2', function () {
+            const migrate = migrations.find((m) => m.version === 2);
+            const migratedSchema = migrate.up(VERSION_1);
+            expect(migratedSchema).to.eql(VERSION_2);
+        });
+    });
     it('should update an undefined version to current version', function () {
         const migratedSchema = migrate(VERSION_UNDEFINED);
-        expect(migratedSchema).to.eql(VERSION_1);
+        expect(migratedSchema).to.eql(VERSION_2);
     });
 });