+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.53.1](https://github.com/zachowj/node-red-contrib-home-assistant-websocket/compare/v0.53.0...v0.53.1) (2023-07-12)
+
+
+### Bug Fixes
+
+* **switch:** Change to strict checking of payload for onTrigger ([6c1a795](https://github.com/zachowj/node-red-contrib-home-assistant-websocket/commit/6c1a795ae4fdc14ba28c450764603b4245285b6a)), closes [#983](https://github.com/zachowj/node-red-contrib-home-assistant-websocket/issues/983)
+
+## [0.53.0](https://github.com/zachowj/node-red-contrib-home-assistant-websocket/compare/v0.52.0...v0.53.0) (2023-07-12)
+
+
+### Features
+
+* **webhook:** Add allowed methods to webhooks ([48ebdd8](https://github.com/zachowj/node-red-contrib-home-assistant-websocket/commit/48ebdd8008f56ae3bbd8d7b938a8a0ec53c60149))
+
+
+### Bug Fixes
+
+* **sentence:** Catch error from custom outputs ([867289a](https://github.com/zachowj/node-red-contrib-home-assistant-websocket/commit/867289a6bfb3bdc77adfa1756adc5489f30e6983))
+
+## [0.52.0](https://github.com/zachowj/node-red-contrib-home-assistant-websocket/compare/v0.51.1...v0.52.0) (2023-07-12)
+
+
+### Features
+
+* **sentence:** Add sentence trigger node ([1ebf3a8](https://github.com/zachowj/node-red-contrib-home-assistant-websocket/commit/1ebf3a8d2e6bed004d1573d87a911052f23e46f5))
+
 ## [0.51.1](https://github.com/zachowj/node-red-contrib-home-assistant-websocket/compare/v0.51.0...v0.51.1) (2023-07-06)